+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.2 - 2023-11-11
+### Added
+- Auto-fetch duplicates again when reaching the end of the list
+### Removed
+- Nextcloud 26 is no longer supported (DuplicateFinder will always be updated for the 2 last versions)
+
 ## 1.1.1 - 2023-11-10
 ### Fixed
 - Fix [#24](https://github.com/eldertek/duplicatefinder/issues/24)